+++ src/Exceptions/EntropyException.php
@@ -1,0 +1,9 @@
+<?php
+namespace JoelShepherd\CreateWith\Exceptions;
+
+use RuntimeException;
+
+class EntropyException extends RuntimeException
+{
+    //
+}